% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_3_matrices}
\alias{arma_3_matrices}
\title{Outer products multiplied by different integers (1, ..., 3)}
\usage{
arma_3_matrices(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Outer products multiplied by different integers (1, ..., 3)
}
